cosmari/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf_yasg==1.21.8
elasticsearch==8.17.1
python-dotenv==1.0.1
Requests==2.32.3
django-cors-headers